fix server crash problem

+++ env_set.sh
@@ -5,34 +5,45 @@
 # sudo apt-get -f install build-essential pkg-config
 # go get gopkg.in/confluentinc/confluent-kafka-go.v1/kafka
 
-MYSQL_INSTALL_GROUPNAME=mysql
-MYSQL_INSTALL_USERNAME=zxqiu
-SYSTEM_USERNAME=zxqiu
-CODE_PATH=$(pwd)
-MYSQL_PATH=/usr/local
 
-DATABASE_USER_NAME=root
-DATABASE_USER_PASSWD=root
+# setup mysql on ubuntu
+
+MYSQL_INSTALL_GROUPNAME=mysql
+MYSQL_INSTALL_USERNAME=mysql
+SERVICE_USERNAME=zxqiu
+SERVICE_PWD=12345
+
+sudo apt install mysql-server -y
 
 # setup configurations
-echo "[mysqld]" > my.cnf
-echo "datadir="$MYSQL_PATH"/mysql/data" >> my.cnf
-echo "socket=/tmp/mysql.sock" >> my.cnf
-echo "symbolic-links=0" >> my.cnf
-echo "character-set-server=utf8" >> my.cnf
-echo "collation-server=utf8_general_ci" >> my.cnf
-echo "[mysqld_safe]" >> my.cnf
-sudo mv my.cnf /etc/my.cnf
+if test -f "/etc/my.cnf"; then
+  echo "/etc/my.cnf already exist. It will not be updated"
+else
+  echo "creating new /etc/my.cnf"
+  echo "[mysqld]" > my.cnf
+  echo "datadir=/var/lib/mysql" >> my.cnf
+  echo "socket=/tmp/mysql.sock" >> my.cnf
+  echo "symbolic-links=0" >> my.cnf
+  echo "character-set-server=utf8" >> my.cnf
+  echo "collation-server=utf8_general_ci" >> my.cnf
+  echo "[mysqld_safe]" >> my.cnf
+  sudo mv my.cnf /etc/my.cnf
+fi
 
-echo "export PATH="$(pwd)"/bin/:$PATH" > mysql.sh
-sudo mv mysql.sh /etc/profile.d/
-source /etc/profile.d/mysql.sh
+# there might be some issues with these folders. fix them first.
+sudo chown mysql:mysql /var/log/mysql
+sudo chown mysql:mysql /var/lib/mysql
+sudo chmod 750 /var/lib/mysql
 
-# start mysql server and init root user
-killall mysqld
-support-files/mysql.server start
-bin/mysqladmin -u $DATABASE_USER_NAME password $DATABASE_USER_PASSWD
-sudo cp support-files/mysql.server /etc/init.d/mysql.server
 
-# create wallet database for project
-bin/mysql -u $DATABASE_USER_NAME --password=$DATABASE_USER_PASSWD --execute="create database wallet"
+# initialize and start service
+sudo mysql --initialize-insecure --user=mysql
+sudo systemctl restart mysql.service
+
+# create user zxqiu
+mysql -u root --execute="ALTER USER 'root'@'localhost' IDENTIFIED BY 'root'"
+mysql -u root --password="root" --execute="CREATE USER '$SERVICE_USERNAME'@'localhost' IDENTIFIED BY '$SERVICE_PWD'"
+mysql -u root --password="root" --execute="GRANT ALL PRIVILAGES ON *.* TO '$SERVICE_USERNAME'@'localhost'"
+
+# create db
+mysql -u $SERVICE_USERNAME --password="$SERVICE_PWD" --execute="CREATE DATABASE payment_service_db"
